+++ lib/capito/translation.rb
@@ -9,5 +9,14 @@ module Capito
     def locale
       read_attribute(:locale).try(:to_sym)
     end
+
+    private
+
+    def destroy_model_without_translation
+      model = translated_model
+      if model && !model.destroyed? && !model.will_destroy? && (model.translations - [self]).empty?
+        model.destroy
+      end
+    end
   end
 end